refactor: 경로 수정 및 id 통일성 부여
