+++ Mytube/src/main/java/com/flab/Mytube/controller/MovieController.java
@@ -1,8 +1,8 @@
 package com.flab.Mytube.controller;
 
-import com.flab.Mytube.movie.request.InsertMovieRequest;
-import com.flab.Mytube.movie.request.InsertPostRequest;
-import com.flab.Mytube.movie.request.JoinChatRequest;
+import com.flab.Mytube.dto.movie.request.InsertMovieRequest;
+import com.flab.Mytube.dto.movie.request.InsertPostRequest;
+import com.flab.Mytube.dto.movie.request.JoinChatRequest;
 import com.flab.Mytube.service.MovieService;
 import com.flab.Mytube.service.StreamingService;
 import lombok.RequiredArgsConstructor;
@@ -21,18 +21,18 @@ public class MovieController {
     public BigInteger reserve(@RequestBody InsertPostRequest.Param param){
         // 관련 dto 생성하여 매개변수로 전달
         // live 예약하기
-        return streamingService.reserveMovie(param).getID();
+        return streamingService.reserveMovie(param).getId();
     }
     @PostMapping("/upload")
     public long upload(@RequestBody InsertMovieRequest param){
         // 동영상 업로드
-        return movieService.insertMovie(param).getID();
+        return movieService.insertMovie(param).getId();
     }
 
     @PostMapping("/{movie_id}/chat")
     public BigInteger joinChat(@RequestBody JoinChatRequest param, @PathVariable("movie_id") BigInteger movie_id){
         //라이브 채팅 참여
-        return streamingService.joinChat(param, movie_id).getID();
+        return streamingService.joinChat(param, movie_id).getId();
     }
 
     @PostMapping("/{movie_id}/store")
