+++ Mytube/src/main/java/com/flab/Mytube/dto/movie/request/InsertPostRequest.java
@@ -1,4 +1,4 @@
-package com.flab.Mytube.movie.request;
+package com.flab.Mytube.dto.movie.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
