+++ Mytube/src/main/java/com/flab/Mytube/dto/movie/response/InsertMovieResponse.java
@@ -1,16 +1,14 @@
-package com.flab.Mytube.movie.response;
+package com.flab.Mytube.dto.movie.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigInteger;
-
 @Getter
 @Setter
 @AllArgsConstructor
 public class InsertMovieResponse {
-    private long ID;
+    private long id;
     private int code;
     private String message;
 }
