+++ Mytube/src/main/java/com/flab/Mytube/dto/movie/response/InsertPostResponse.java
@@ -1,4 +1,4 @@
-package com.flab.Mytube.movie.response;
+package com.flab.Mytube.dto.movie.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,8 +7,8 @@ import java.math.BigInteger;
 
 @Getter
 @AllArgsConstructor
-public class JoinChatResponse {
-    private BigInteger ID;
+public class InsertPostResponse {
+    private BigInteger id;
     private int code;
     private String message;
 }
