+++ Mytube/src/main/java/com/flab/Mytube/dto/movie/response/JoinChatResponse.java
@@ -1,15 +1,14 @@
-package com.flab.Mytube.movie.response;
+package com.flab.Mytube.dto.movie.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.RequiredArgsConstructor;
 
 import java.math.BigInteger;
 
 @Getter
 @AllArgsConstructor
-public class InsertPostResponse {
-    private BigInteger ID;
+public class JoinChatResponse {
+    private BigInteger id;
     private int code;
     private String message;
 }
