+++ Mytube/src/main/java/com/flab/Mytube/mapper/MovieMapper.java
@@ -1,6 +1,6 @@
 package com.flab.Mytube.mapper;
 
-import com.flab.Mytube.movie.request.InsertMovieRequest;
+import com.flab.Mytube.dto.movie.request.InsertMovieRequest;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 import org.springframework.stereotype.Component;
