+++ Mytube/src/main/java/com/flab/Mytube/service/MovieService.java
@@ -1,7 +1,7 @@
 package com.flab.Mytube.service;
 
-import com.flab.Mytube.movie.request.InsertMovieRequest;
-import com.flab.Mytube.movie.response.InsertMovieResponse;
+import com.flab.Mytube.dto.movie.request.InsertMovieRequest;
+import com.flab.Mytube.dto.movie.response.InsertMovieResponse;
 import com.flab.Mytube.mapper.MovieMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -16,6 +16,6 @@ public class MovieService {
     @Transactional
     public InsertMovieResponse insertMovie(InsertMovieRequest param){
         long resultID = movieMapper.addMovie(param);
-        return new InsertMovieResponse(param.getID(), 201, "success!");
+        return new InsertMovieResponse(param.getId(), 201, "success!");
     }
 }
