+++ Mytube/src/main/resources/mybatis/mapper/MovieMapper.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.flab.Mytube.mapper.MovieMapper">
-    <insert id="addMovie" useGeneratedKeys="true" keyProperty="ID" parameterType="com.flab.Mytube.movie.request.InsertMovieRequest">
+    <insert id="addMovie" useGeneratedKeys="true" keyProperty="ID" parameterType="com.flab.Mytube.dto.movie.request.InsertMovieRequest">
         INSERT INTO
         movie (USER_ID, SUBJECT, URL)
         VALUES
